First Commit.
Adding a new ads scraper.

It is based on scraping adsense ads. It currently focuses on two types of ads and is scraped only from the site geeksforgeeks.org, as they have a particular structure of embedding google ad frames in their webpage which is being exploited here.

The following structure is currently been observed in the google ads in the site geeksforgeeks.org:
<ins class='adsbygoogle'>
    <ins>
        <ins>
            <iframe>
                ...
                    <iframe id='google_ads_frame..'>

There are two types of ads that are being embedded here:
1. Ad with only one image.
2. Ad that contain iframe within it. Saving the innerhtml of these iframe will contain the actual google ad html code.

Saving these two types of ads and removing all script tags to further clean it.