Changes in code structure. Now more Robust.
Currently looking into only 3 types of ads:
1. Embedded ads: These will have iframe within the ad_iframe with a src
attribute.
2. Display ads: Simple image ads that will be inside the ad_iframe with
a <a> tag and href along with it. Inside this, there will be a img with
src to the image ad
3. A HTML ad that is simply there. This will contain a <a> tag with href
to the place where the ad is redirected.
TODO: 3rd type ad is to be optimized, as many a times this type of ad
was unable to find a <a> tag.